Replace my-app with main content
