+++ src/App.test.tsx
@@ -1,9 +1,17 @@
+import { MathJaxContext } from 'better-react-mathjax';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 
 import App from './App';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  const root = ReactDOM.createRoot(div);
+  root.render(
+    <React.StrictMode>
+      <MathJaxContext>
+        <App />
+      </MathJaxContext>
+    </React.StrictMode>,
+  );
 });
