+++ src/index.tsx
@@ -1,3 +1,4 @@
+import { MathJaxContext } from 'better-react-mathjax';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
@@ -8,7 +9,9 @@ const root = ReactDOM.createRoot(
 );
 root.render(
   <React.StrictMode>
-    <App />
+    <MathJaxContext>
+      <App />
+    </MathJaxContext>
   </React.StrictMode>,
 );
